feat(tests): test for 1 query
