--- wiki-1.0/tests/data-index.csv
@@ -1,100 +1,0 @@
-Randy[SEP]I wish I could!
-Chef[SEP]OOH
-Cartman[SEP]How about we have us
-KurtRussell[SEP]They're raping us and it huuurts!
-Cartman[SEP]This way
-Kyle[SEP]I'm not happy with the direction this company is taking.
-Butler[SEP]Here he is.
-Clyde[SEP]Uh.
-Stan[SEP]That's my silly grampa.
-Sharon[SEP]Sorry
-KyleBroflovski[SEP]OooOoOoOooOOO!
-Coon[SEP]I work on my own now.
-LeadBiker[SEP]I'm trying to talk
-MrDawkins[SEP]I'll thell them.
-Marvin[SEP]Who's Stan?
-Stockbroker[SEP]Just talk to the Treasury Department in DC.
-Jimbo[SEP]Christ
-Stan[SEP]Hey
-Charlie[SEP]Of next year.
-Kyle[SEP]Nothing's changed.
-EmoGirl[SEP]I'd rather kill myself than put on gym shorts.
-Stan[SEP]I just wanna be a normal kid again.
-Terrance[SEP]I was an asshole
-Teresa[SEP]What a stupid bitch!
-Jimmy[SEP]I just think it's a fabulous app
-AlecBaldwin[SEP]Some little faggot in my head.
-MrMackey[SEP]Freely?
-Fiancée[SEP]I'll pilot!
-Man[SEP]Allб
-Cartman[SEP]Gahaaa!
-Kelly[SEP]Come on
-Clyde[SEP]She's totally staring you down.
-Butters[SEP]Leave me alone!
-Stan[SEP]We don't have any weapons.
-Cartman[SEP]I can hear Kyle
-LeadBiker[SEP]Ju-just ignore him!
-Cartman[SEP]I'm going to find out where it is!
-Kyle[SEP]Then we'll wait
-Kyle[SEP]Oh
-Wendy[SEP]Mayor
-Jimbo[SEP]Surrender your men
-Stan[SEP]All right Avengers
-Rod[SEP]Oh.
-Butters[SEP]Ah-I'm grounded for lookin' stupid in my school picture.
-Bebe[SEP]They called me a slut
-Postum[SEP]Oh
-Cartman[SEP]I can't-
-Scout[SEP]Kid
-MrsGarrison[SEP]My eyes are opened!
-Cartman[SEP]Dude
-PresidentWulff[SEP]Ich werde zuerst umbringen!
-Randy[SEP]To be honest
-Linda[SEP]Why do you children live in this town alone?
-Cartman[SEP]It's the only way.
-Cartman[SEP]Age
-Kyle[SEP]Timmy saw the turkey and wanted to get it.
-Cartman[SEP]Get a hold of yourself
-Cartman[SEP]Fuck you Bebe!
-Cartman[SEP]You serve
-Pablo[SEP]Quite an amazing creature.
-Man10[SEP]'Scuse me
-Jimbo[SEP]Ned
-Kyle[SEP]Quit it
-Homeowner[SEP]They're gonna betray you at the Red Robin wedding!
-PrincipalVictoria[SEP]Listen Eric!
-Chef[SEP]Oh
-Kyle[SEP]Yes
-Randy[SEP]Yuh
-Lucas[SEP]Don't say anything
-Craig[SEP]Nope.
-MrGarrison[SEP]Jimbo
-Kyle[SEP]Because you're a fucking fatass
-Janet[SEP]are Lesbos!
-Cartman[SEP]Yes
-Cartman[SEP]Outlander
-Stan[SEP]Disqualified?
-Man[SEP]Nadia
-ATFlead[SEP]Even if it means we have to kill each and every one of them.
-Liane[SEP]How are you doing?
-Harry[SEP]It was a tough time for all of Canada.
-Cartman[SEP]next week sometime.
-Randy[SEP]Yes
-Cartman[SEP]It's simple science
-Katherine[SEP]Could you get that
-Producer[SEP]Stallone thinks your singer has a lot of talent.
-Sharon[SEP]Stanley
-Stephen[SEP]Look
-Cartman[SEP]Ringy-ding?
-Cartman[SEP]AAAAAA!
-JimRome[SEP]Check that: I'll watch that game.
-Cartman[SEP]Nice
-Gerald[SEP]get out of the house?
-SgtYates[SEP]You boys say you can identify the Red Box thieves?
-Aussie[SEP]Come back here
-ABritishPerson[SEP]But every single day Pip's love for Estella grew.
-Stan[SEP]I hate that toy!
-Man1[SEP]Uh
-Doorman[SEP]Come in.
-CheesyPoofCartman[SEP]Lame.
-Stan[SEP]persecrute our religious beliefs.
